Move arg-handling logic to separate file
